refactor: Group related extension and methods into their own separate files
