+++ lib/shared/constants.dart
@@ -3,9 +3,3 @@ import 'package:flutter/material.dart';
 const kBackgroundColor = Color(0xFF051139);
 const kCardColor = Color(0xFF111B43);
 const kPrimaryColor = Color(0xFF4AD0EE);
-
-extension StringExtension on String {
-  String capitalizeFirstLetter() {
-    return "${this[0].toUpperCase()}${substring(1).toLowerCase()}";
-  }
-}
