+++ lib/shared/string_utils.dart
@@ -1,0 +1,5 @@
+extension StringExtension on String {
+  String capitalizeFirstLetter() {
+    return "${this[0].toUpperCase()}${substring(1).toLowerCase()}";
+  }
+}
